requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
einops==0.8.2
face-alignment==1.4.1
facexlib==0.3.0
filelock==3.19.1
filterpy==1.4.5
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
ftfy==6.3.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime-gpu==1.19.2
opencv-python==4.13.0.92
packaging==26.0
pillow==11.3.0
protobuf==6.33.5
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-vlc==3.0.21203
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.14.0
tifffile==2024.8.30
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
zipp==3.23.0
